requirements.txt
absl-py==2.3.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.3.0
av==14.4.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
chex==0.1.89
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
distro==1.9.0
docstring_parser==0.16
et_xmlfile==2.0.0
etils==1.12.2
eval_type_backport==0.2.2
exceptiongroup==1.3.0
filelock==3.18.0
flatbuffers==25.2.10
flax==0.10.6
frozenlist==1.6.2
fsspec==2025.5.1
gast==0.6.0
google-pasta==0.2.0
grpcio==1.73.0
h11==0.16.0
h5py==3.14.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.32.4
humanfriendly==10.0
humanize==4.12.3
idna==3.10
importlib_resources==6.5.2
jax==0.6.1
jaxlib==0.6.1
Jinja2==3.1.6
jiter==0.10.0
keras==3.10.0
libclang==18.1.1
livekit==1.0.8
livekit-agents==1.0.23
livekit-api==1.0.2
livekit-plugins-cartesia==1.0.23
livekit-plugins-deepgram==1.0.23
livekit-plugins-elevenlabs==1.0.23
livekit-plugins-groq==1.0.23
livekit-plugins-noise-cancellation==0.2.4
livekit-plugins-openai==1.0.23
livekit-plugins-silero==1.0.23
livekit-plugins-turn-detector==1.0.23
livekit-protocol==1.0.3
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml_dtypes==0.5.1
mpmath==1.3.0
msgpack==1.1.0
multidict==6.4.4
namex==0.1.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.1.3
onnxruntime==1.22.0
openai==1.84.0
openpyxl==3.1.5
opt_einsum==3.4.0
optax==0.2.4
optree==0.16.0
orbax-checkpoint==0.11.13
packaging==25.0
pandas==2.3.0
pillow==11.2.1
propcache==0.3.1
protobuf==5.29.5
psutil==7.0.0
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.19.1
PyJWT==2.10.1
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
rich==14.0.0
safetensors==0.5.3
scipy==1.15.3
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.2
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.31.0
tensorstore==0.1.75
termcolor==3.1.0
tokenizers==0.21.1
toolz==1.0.0
torch==2.7.1
tqdm==4.67.1
transformers==4.52.4
treescope==0.1.9
types-protobuf==4.25.0.20240417
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
uuid==1.30
watchfiles==1.0.5
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.2
yarl==1.20.0
zipp==3.23.0
